Get host/ip from env.
This is in addition to https://github.com/dmlc/dmlc-core/pull/110.

The host/ip got from network adaptor info in network_util.h is not necessarily reachable from outside the node/container wHen the networks are unidentical or malformed configured. We can specify a host/ip that is guaranteed to be accessible from outside the container/node which is typically provided by the cluster management module like in https://github.com/dmlc/dmlc-core/pull/110.

If the env variable "DMLC_NODE_HOST" is not specified, the program falls back to the previous implementation of querying network interface and getting the ip.